requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
coverage==6.4.2
Django==4.0.6
djangorestframework==3.13.1
freezegun==1.2.1
psycopg2==2.9.3
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
